axi_ethernet: fix dtg warnings
axi_ethernet mdd file for the property axififo-connected is saying that
its a reference value, but the driver tcl, if axi_ethernet is not
connected to a fifo will assign nothing to axififo-connected property.

if there is no value assigned to axififo-connected, but expecting a referenced
value in mdd, dtg is giving the following warnings.
"Only boolean type can have empty value. Fail to add driver(axi_ethernet_0)
property(axififo-connected) type(reference) value()"

So this patch removes the reference in mdd and adds a reference value for the
axififo-connected property, when ethrenet is connected to streaming fifo.

Signed-off-by: Naga Sureshkumar Relli <nagasure@xilinx.com>
Reviewed-by: Kedareswara rao Appana <appanad@xilinx.com>