Fix device tree error while generating bsp
While generating device tree, if interrupt pin of an IP is not connected
then instead of bsp error, display a warning message. while updating ZynqMP
device tree support, proc gen_interrupt_property got changed and in that proc
we are running some hsi commands even though there is no intc, this is
causing the bsp error. this patch fixes that.

Signed-off-by: Naga Sureshkumar Relli <nagasure@xilinx.com>
Acked-by: Kedareswara rao Appana <appanad@xilinx.com>