etherent: Remove compatible property for phy
Compataible property for the phy is not required
As the probing of the phy driver depends on the
PHY vendor id/device id registers.

With the recent changes in the linux framework
if a phy device contains a compatible string other than
standard ethernet phy compatible string it will be treated
as an mdio device rather than a phy device.

Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Acked-by: Punnaiah Choudary Kalluri<punnaia@xilinx.com>