fix error while parsing system.mss file
When using gpio as interrupt controller, we have updated
"#interrupt-cells" in gpio node using add_new_property api, so it adds
that property with #symbol as a comment in mss file. so at the time of
parsing the mss file, sdk throughs error. so change add_new_property api
to add_new_dts_param, so the resultant dts parameter will be added to
dts entry but not to mss file.

Signed-off-by: Naga Sureshkumar Relli <nagasure@xilinx.com>
Reported by: Francesco Ricci <flr@xiphos.com>
Acked-by: Michal Simek <michals@xilinx.com>