axi_dma: Add xlnx,include-dre property in the dma node
In the driver fixed issues with DRE with this commit:
https://gitenterprise.xilinx.com/Linux/linux-xlnx/commit/99d053bb0e87207e0177f8ac236477e812857aee
Driver expects xlnx,include-dre property for DRE
based design's.
Without this property performance will be less for
DRE based design's.

Currently axi_dma tcl is exporting this property
only for 1G based designs.
it is not exporting include-dre property for
10G based design.

This patch fixes this issue.

Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Acked-by: Michal Simek <michals@xilinx.com>