Fix the axi_dma hang issue
In the function add_dma_coherent_prop based on the number of
master interfaces will be looping and check the interface
port is S_AXI_ACP then we add the dma-coherent property.
But is some of the designs the master interfaces can be
NULL, so add the check to fix this case.

Signed-off-by: Venkatesh Yadav Abbarapu <venkatesh.abbarapu@xilinx.com>