Fix the parsing error when using with SDK
In dp.tcl we added tab for the property phys as below
phys = <&lane1 5 0 3 27000000>,
        <&lane0 5 1 3 27000000>;
but sdk throws syntax errors if we use tabs, the notation is fine with
hsi. So removing the tabs which resolves these errors and the phys
property looks like
phys = <&lane1 5 0 3 27000000>, <&lane0 5 1 3 27000000>;
Similarly we added tabs for the clocks fixing the same.

Signed-off-by: Venkatesh Yadav Abbarapu <venkatesh.abbarapu@xilinx.com>