Fix the error when the design has two similar video pipelines
There are some cases where the design consists of multiple similar
video pipelines and this can cause an error because of harcoded port
references. Fixing this to not generate an error instead warn saying
there are multiple similar video pipelines and user need to update
the input and output ports.

Signed-off-by: Venkatesh Yadav Abbarapu <venkatesh.abbarapu@xilinx.com>