mixer: synchronize the video formats
Synchronize with the format descriptions from v_mix_v5_0.tcl.

Fixed ones are,
For stream interface, 5 {RGBA} 6 { YUVA 4:4:4}
For memory interface, 14 {YUVA8} 17 {RGB565}

The naming starts with components in LSB while it's reverse in fourcc, so
- 5 RGBA to be AB24
- 6 YUVA to be AVUY
- 14 YUVA8 to be AVUY
- 17 RGB565 to be BG16

Fourcc is not available for some 10 bit formats, so skipping 10 bit format
handling for those. And some fourccs are duplicates as both memory and
stream formats use same fourcc.

This change is done by comparison in code, but not tested with physical
format validation.

Reported-by: Kenneth Sloat <ksloat@designlinxhs.com>
Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Cc: Vishal Sagar <vishal.sagar@xilinx.com>
Reviewed-by: Venkatesh Yadav Abbarapu <venkatesh.abbarapu@xilinx.com>