common_proc.tcl: soc boot flow enablement
This changes are required to support the classis soc v2 architecture.
In V2 architecture in single xsa we will get pl.pdi and boot pdi.
In regular dfx and classic soc v1 our designs are block container designs.
But in classic SOC v2 the designs are flat designs and we can not use the
existing infra. The firmname-name(pdi anme) property is required  for
libdfx to map the dtbo and pdi. In v2 using the new get_hw_files hsi api
getting the pdi name.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>
Acked-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>