scugic.tcl:enabling gic_its node status
gic_its node status was disabled default in board dtsi file.
This should be enabled if gic_its available.
To enable the status there is no hsi property earlier.
Now HSI team added a property based on that enabling the status
and checking for cpm enablement also as this is only required
for cpm in versal.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>

Acked-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>