common_proc.tcl:removing sorting for custom dts inclusion
When user enables "pl custom" or "pl partial custom" in DTG
while including the dtsi files in user specified dts file
we are ordering based on name which is casuing issues if
we have case sensetive rprm names.

Actually ordering is not required as while including in dts
file itself we are taking care of the order.
In this patch removing the ordering.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>
Acked-by: Venkatesh Yadav Abbarapu <venkatesh.abbarapu@amd.com>