common_proc.tcl:input endpoint missing for tpg ip
If tpg is getting input from gamma ip then setting input
ip to gamma as the get_in_connect_ip API is traversing through
first input source and which might not be correct as input/output
endpoints should be for immediate mem mapped ips.
Corrected for gamma ip if it is input to tpg.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>
Acked-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>