vid_phy_ctrl:hdmi_gt_ctrl: Generate Channels only if connected to DP/HDMI
vid_phy_controller is generating the channel nodes based on the
tx/rx channels count. Check if channel pin is connected or not
and check memory mapped or not then generate the channel nodes.

Add pinname + connected IP instances as node name and
vid IP instance + txphy_lane0/1/2/3 as label.
This will help when multiple phy_controllers are present
in design it will generate the nodes with unique names.

Signed-off-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>
Acked-by: Varalaxmi Bingi <Varalaxmi.bingi@amd.com>