axi_ethernet.tcl: Fix the issue when dclk clock index is 0
gen_clk_property api is not adding the & if clock index is 0 because
the add_dt_propety api will add it in creation of clocks property.
For axi_ethernet when dclk clock index 0 and adds at middle of clocks
property causing the build issue.
Fix this issue by checking the dclk startswith & or not and add it, if
not has.

Signed-off-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>
Acked-by: Swagath Gadde <swagath.gadde@amd.com>