common_proc.tcl: pass one intc instance
When there are multiple instances of intc in MB
design then we passing all. For platforms we are passing
one by one. The check to pass one by one missed for MB
adding that check as part of this patch.
In Vitis gen_property is checking properly now
about passed object.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>
Acked-by: Swagath Gadde <swagath.gadde@amd.com>