common_proc.tcl: generating the fixed-factor-clock node
Generating fixed-factor-clock instead of fixed-clock
to skip the fclk driver.
Generating fixed-factor-clock instead of fixed-clock to skip the fclk driver.
fixed-factor-clock requires clock-div and clock-mult and those we
are not having in vivado or hsi.
We are getting parent clock frequency and pin clock from those
generating clock-div and clock-mult.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>
Acked-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>