common_proc.tcl: enabling cpm4 and cpm5 node status
The versal.dtsi cpm base address was not matching with the one in design.
As we cross verify the base address with versal.dtsi while updating the status, status is not setting.
Actually driver team should fix this in versal.dtsi but it will break the backward compatibility.
To avoid that reading the SLR address from design and updating back.
But in CPM4 and CPM5 designs we have both SLR address properties.
and always setting cpm5 slr address for both and status okay not setting for cpm4.
To differentiate whether it was cpm4 or cpm5 we have one property in "versal_cips_0_cpm_0_psv_cpm" but
this is not memory mapped so can not be listed under hsi get_drivers.
So using hsi get_cells to get the above ip and reading the revision number
and based on the revision updating the corresponding SLR addresses.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>
Acked-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>