arm64: versal: Describe EEPROM on i2c on vek280-revA/vhk158-revA/vpk180-revA
i2c bus is shared between DUT and SC(System Controller) and it is not
recommended to use any device connected to i2c mux. In case of EEPROM which
is placed on i2c bus (without i2c mux) it shouldn't be a problem to
describe it. EEPROM is read only that's why it shouldn't be a problem to
access it via i2c multimaster bus from both sides.

Similar change has been done by commit e060d9d6f116 ("arm64: versal:
Describe EEPROM on i2c on vek280"). vpk120-revA and
versal-vp-x-a2785-00-revA are not updated because i2c eeprom is behing i2c
mux.

vpk180 is also updated to be aligned with others.

Signed-off-by: Michal Simek <michal.simek@amd.com>
Message-ID: <bf0547f2bc44ac514c59c19a50c6c337015b874e.1733225984.git.michal.simek@amd.com>
Acked-by: Ashwini Lomate <ashwini.lomate@amd.com>