axi_mcdma.tcl: fixing xlnx, addrwidth property
The xlnx,addrwidth property is not compiled properly
with dtc when we have xlnx,addrwidth = /bits/ 8 <0x20>
it was generating as xlnx,addrwidth = [20] after compilation
and driver is unable to find the property.

The dtc expecting it should be as xlnx,addrwidth = /bits/ 8 <0x0 0x0 0x0 0x20>
then the property is getting generated properly as xlnx,addrwidth = <0x20>
and driver is able to detect the property properly.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>
Acked-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>