arm64: zynqmp: Remove remoteproc nodes
Both of nodes are using the same address which is reported as error and
none of them should be enabled by default. Remove them from zynqmp.dtsi
which is the first step and likely content should be moved to own dtsi
which can be included when this configuration is requested.

Signed-off-by: Michal Simek <michal.simek@amd.com>
State: pending

Acked-for-series: Varalaxmi Bingi <Varalaxmi.bingi@amd.com>