dp_rx: Add Support for multiple DP Rx instances
If there are multiple vid_edid (and dp_rx) instances in a design,
the DTG cannot figure out which vid_edid to use to complete the
DT properties of the connected dp_rx.

To create a recognizable bond between a vid_edid and
the connected dp_rx, they can be placed in the same hierarchical subblock.
This way, their names get common prefixes.

The dp_rx script is updated to recognize such a relationship
and use the vid_edid whose name most closely matches the dp_rx name.

This also prevents crashes that occurred in previous versions,
when multiple vid_edid were present in a single design.

Signed-off-by: Eddy Debaere <eddy.debaere.ext@barco.com>
Signed-off-by: Lakshmi Prasanna Eachuri <lakshmi.prasanna.eachuri@amd.com>

Acked-for-series: Gaddipati, Naveen <naveen.gaddipati@amd.com>